@@ -5,61 +5,62 @@
     "description": "small web app to save points for players of the german card game 'Doppelkopf'",
     "keywords": ["symfony", "doppelkopf"],
     "require": {
-        "php": "^7.3.0",
+        "php": "^7.4 || ^8.0",
         "ext-iconv": "*",
-        "doctrine/collections": "1.6.7",
-        "doctrine/dbal": "2.11.1",
-        "doctrine/doctrine-bundle": "2.1.2",
-        "doctrine/doctrine-migrations-bundle": "2.2.0",
-        "doctrine/orm": "2.7.3",
-        "knplabs/knp-paginator-bundle": "5.3.0",
-        "knplabs/knp-components": "2.4.2",
-        "symfony/asset": "5.1.6",
-        "symfony/config": "5.1.6",
-        "symfony/console": "5.1.6",
-        "symfony/dependency-injection": "5.1.6",
-        "symfony/doctrine-bridge": "5.1.6",
-        "symfony/dotenv": "5.1.6",
-        "symfony/flex": "1.9.4",
-        "symfony/form": "5.1.6",
-        "symfony/framework-bundle": "5.1.6",
-        "symfony/http-foundation": "5.1.6",
-        "symfony/http-kernel": "5.1.6",
-        "symfony/routing": "5.1.6",
-        "symfony/translation": "5.1.6",
-        "symfony/translation-contracts": "2.2.0",
-        "symfony/twig-bundle": "5.1.6",
-        "symfony/validator": "5.1.6",
-        "symfony/yaml": "5.1.6"
+        "doctrine/collections": "~1.8.0",
+        "doctrine/dbal": "~2.13.9",
+        "doctrine/doctrine-bundle": "~2.7.0",
+        "doctrine/doctrine-migrations-bundle": "~2.2.3",
+        "doctrine/orm": "~2.20.2",
+        "knplabs/knp-components": "~2.5.0",
+        "knplabs/knp-paginator-bundle": "~5.9.0",
+        "symfony/asset": "~5.4.0",
+        "symfony/config": "~5.4.0",
+        "symfony/console": "~5.4.0",
+        "symfony/dependency-injection": "~5.4.0",
+        "symfony/doctrine-bridge": "~5.4.0",
+        "symfony/dotenv": "~5.4.0",
+        "symfony/flex": "~1.22.0",
+        "symfony/form": "~5.4.0",
+        "symfony/framework-bundle": "~5.4.0",
+        "symfony/http-foundation": "~5.4.0",
+        "symfony/http-kernel": "~5.4.0",
+        "symfony/routing": "~5.4.0",
+        "symfony/runtime": "~5.4.0",
+        "symfony/translation": "~5.4.0",
+        "symfony/translation-contracts": "~2.5.4",
+        "symfony/twig-bundle": "~5.4.0",
+        "symfony/validator": "~5.4.0",
+        "symfony/yaml": "~5.4.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "2.16.4",
-        "jakub-onderka/php-parallel-lint": "1.0.0",
-        "jakub-onderka/php-var-dump-check": "0.3.0",
-        "kubawerlos/php-cs-fixer-custom-fixers": "2.3.0",
-        "phpstan/extension-installer": "1.0.5",
-        "phpstan/phpstan": "0.12.48",
-        "phpstan/phpstan-doctrine": "0.12.20",
-        "phpstan/phpstan-phpunit": "0.12.16",
-        "phpstan/phpstan-symfony": "0.12.8",
-        "phpunit/phpunit": "9.4.0",
-        "psalm/plugin-symfony": "1.4.4",
-        "sensiolabs/security-checker": "^6.0",
-        "symfony/browser-kit": "5.1.6",
-        "symfony/css-selector": "5.1.6",
-        "symfony/error-handler": "5.1.6",
-        "symfony/phpunit-bridge": "5.1.6",
-        "symfony/stopwatch": "5.1.6",
-        "symfony/web-profiler-bundle": "5.1.6",
-        "symplify/coding-standard": "8.3.26",
-        "symplify/easy-coding-standard": "8.3.26",
-        "vimeo/psalm": "3.16"
+        "friendsofphp/php-cs-fixer": "3.75.0",
+        "kubawerlos/php-cs-fixer-custom-fixers": "3.24.0",
+        "phpstan/extension-installer": "1.4.3",
+        "phpstan/phpstan": "2.1.12",
+        "phpstan/phpstan-deprecation-rules": "2.0.1",
+        "phpstan/phpstan-doctrine": "2.0.2",
+        "phpstan/phpstan-phpunit": "2.0.6",
+        "phpstan/phpstan-symfony": "2.0.4",
+        "phpunit/phpunit": "~9.6.22",
+        "symfony/browser-kit": "~5.4.0",
+        "symfony/css-selector": "~5.4.0",
+        "symfony/error-handler": "~5.4.0",
+        "symfony/phpunit-bridge": "~5.4.0",
+        "symfony/stopwatch": "~5.4.0",
+        "symfony/web-profiler-bundle": "~5.4.0"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "phpstan/extension-installer": true,
+            "symfony/flex": true,
+            "symfony/runtime": true
+        }
     },
     "autoload": {
         "psr-4": {
